requirements.txt
sentence-transformers>=2.3.0
faiss-cpu>=1.13.0
transformers>=4.40.0
langchain>=0.2.0
torch>=2.3.0
datasets>=2.14.6
numpy>=2.0.0,<2.3.0
tqdm==4.66.1
imageio>=2.5,<3.0
